Provide list of incompatible components for NSXv plugin
Explicitly enumerate components that cannot be enabled with NSXv plugin.
Otherwise they are marked with tooltip that says "This component was not
tested with the following components: ..."

NSXv plugin is not compatible with:
* Ceph as block and ephemeral storage is not compatible, because it
  cannot be connected to ESXi hosts via vCenter
* Sahara and Murano are incompatible due to deployment errors
* Ironic is not compatible, because it requires Neutron with VLAN
  segmentation support

List qemu as compatible hypervisor, because otherwise message "This
component was not tested... with Neutron NSX plugin" appears may confuse
end user.

List incompatible component in "Limitations" section of user guide.

Cherry-picked from e0ceda1b
Closes-bug: 1550169
Change-Id: I48e388615a68fbc593e912d265523da780243a92